# docs/api/cotsel-dashboard-gateway.openapi.yml
openapi: 3.1.0
info:
  title: Cotsel Dashboard Gateway API
  version: 0.1.0
  summary: Contract-first admin/operator gateway for Cotsel-Dash.
  description: |
    Contract-first API specification for the Cotsel-Dash admin/operator dashboard to interact with Cotsel.

    Discovery outcome:
    - Dedicated in-repo dashboard gateway runtime: `gateway/` Express + TypeScript service.
    - Existing repo auth/session surface: `auth` service issues Bearer session tokens.
    - Existing repo governance source of truth: `contracts/src/AgroasysEscrow.sol`, `sdk/src/modules/adminSDK.ts`, and governance runbooks.
    - Existing repo compliance source of truth: policy/runbooks only; no dedicated compliance HTTP service exists yet.

    This specification is therefore contract-first for the dashboard gateway, but it is grounded in existing repo truth.

    Gateway authorization policy:
    - Auth uses existing auth-service Bearer sessions.
    - Gateway role mapping is derived from current auth role `admin`, which maps to `operator:read` and `operator:write`.
    - All mutation endpoints additionally require `GATEWAY_ENABLE_MUTATIONS=true` and membership in `GATEWAY_WRITE_ALLOWLIST`.
...
servers:
  - url: http://127.0.0.1:3600/api/dashboard-gateway/v1
    description: Local contract-first gateway base URL.
...
security:
  - sessionBearer: []
paths:
...
  /trades:
    get:
      tags: [Trades]
      summary: List dashboard trade records
      description: |
        Indexer-backed dashboard trade registry. This surface is read-only and combines:
        - indexed trade lifecycle state from `indexer/schema.graphql`
        - indexed trade-event history for timeline rendering
        - current gateway compliance ledger status, when available

        Amounts are normalized to display-scale USDC values for dashboard rendering.
      operationId: listTrades
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Trade registry records.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /trades/{tradeId}:
    get:
      tags: [Trades]
      summary: Get trade detail by trade id
      operationId: getTrade
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/TradeId'
      responses:
        '200':
          description: Trade detail record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /treasury:
    get:
      tags: [Treasury]
      summary: Get treasury state and sweep visibility snapshot
      description: |
        Operator read surface for current treasury state. This snapshot is derived from chain RPC
        governance reads and makes claim/sweep visibility explicit for dashboard consumers.
      operationId: getTreasury
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '200':
          description: Treasury snapshot with explicit sweep visibility and freshness metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreasurySnapshotResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /treasury/actions:
    get:
      tags: [Treasury]
      summary: List treasury governance actions for dashboard history views
      description: |
        Treasury history read surface derived from the gateway governance ledger. This route is
        restricted to treasury-related categories and uses deterministic cursor pagination to remain
        consistent with the underlying governance action ledger.
      operationId: listTreasuryActions
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/TreasuryActionCategory'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/GovernanceActionStatus'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Treasury governance history with cursor pagination and freshness metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreasuryActionListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: session-id
      description: |
        Session bearer token issued by the in-repo auth service.

        Production path:
        - trusted upstream session exchange via `/api/auth/v1/session/exchange/agroasys`

        Compatibility path:
        - `/api/auth/v1/login` after explicit wallet-signature login when that legacy flow is enabled

        Repo evidence:
        - `auth/src/api/controller.ts`
        - `auth/src/middleware/middleware.ts`
        - `auth/README.md`

        The gateway must resolve the session subject and enforce endpoint-level role policy. Governance and compliance mutations require operator/admin authorization.
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Client-generated request identifier for end-to-end traceability. Gateway should generate one if absent.
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
      description: Stable correlation identifier linking dashboard actions to downstream service logs, trade events, and incident artifacts.
...
    TradeId:
      name: tradeId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    UnauthorizedError:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Authenticated subject lacks required role or scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        timestamp:
          type: string
          format: date-time
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [AUTH_UNAUTHORIZED, GOV_STATE_CONFLICT, CMP_PROVIDER_UNAVAILABLE]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
        requestId:
          type: string
    EvidenceLink:
      type: object
      required: [kind, uri]
      properties:
        kind:
          type: string
          enum: [runbook, incident, ticket, tx, event, document, log, dashboard, other]
        uri:
          type: string
        note:
          type: string
    MutationAuditInput:
      type: object
      required: [reason, evidenceLinks, ticketRef]
      properties:
        reason:
          type: string
          minLength: 8
          maxLength: 2000
        evidenceLinks:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/EvidenceLink'
        ticketRef:
          type: string
          minLength: 2
          maxLength: 128
    MutationAuditRecord:
      allOf:
        - $ref: '#/components/schemas/MutationAuditInput'
        - type: object
          required: [actorSessionId, actorWallet, actorRole, createdAt, requestedBy]
          properties:
            actorSessionId:
              type: string
            actorAccountId:
              type: string
              nullable: true
            actorWallet:
              type: string
            actorRole:
              type: string
            createdAt:
              type: string
              format: date-time
            requestedBy:
              type: string
            approvedBy:
              type: array
              items:
                type: string
            finalSignerWallet:
              type: string
              nullable: true
            finalSignerVerifiedAt:
              type: string
              format: date-time
              nullable: true
...
    TradeStatus:
      type: string
      enum: [locked, stage_1, stage_2, completed, disputed]
    TradeComplianceStatus:
      type: string
      enum: [pass, fail, unavailable]
    TradeTimelineEvent:
      type: object
      required: [stage, timestamp, actor]
      properties:
        stage:
          type: string
        timestamp:
          type: string
          format: date-time
        actor:
          type: string
        txHash:
          type: string
          nullable: true
          description: Canonical EVM transaction hash when available.
        explorerUrl:
          type: string
          nullable: true
          description: Canonical chain explorer link derived from `txHash` when available.
        detail:
          type: string
    Trade:
      type: object
      required:
        - id
        - buyer
        - supplier
        - amount
        - currency
        - status
        - createdAt
        - updatedAt
        - ricardianHash
        - platformFee
        - logisticsAmount
        - timeline
        - complianceStatus
        - settlement
      properties:
        id:
          type: string
        buyer:
          type: string
        supplier:
          type: string
        amount:
          type: number
        currency:
          type: string
          const: USDC
        status:
          $ref: '#/components/schemas/TradeStatus'
        txHash:
          type: string
          nullable: true
          description: Canonical EVM transaction hash for the trade lock when available.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        ricardianHash:
          type: string
        platformFee:
          type: number
        logisticsAmount:
          type: number
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TradeTimelineEvent'
        complianceStatus:
          $ref: '#/components/schemas/TradeComplianceStatus'
        settlement:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/TradeSettlementState'
    TradeSettlementState:
      type: object
      required:
        - handoffId
        - platformId
        - platformHandoffId
        - phase
        - settlementChannel
        - displayCurrency
        - displayAmount
        - executionStatus
        - reconciliationStatus
        - callbackStatus
        - providerStatus
        - txHash
        - externalReference
        - latestEventType
        - latestEventDetail
        - latestEventAt
        - callbackDeliveredAt
        - createdAt
        - updatedAt
      properties:
        handoffId:
          type: string
        platformId:
          type: string
        platformHandoffId:
          type: string
        phase:
          type: string
        settlementChannel:
          type: string
        displayCurrency:
          type: string
        displayAmount:
          type: number
        executionStatus:
          $ref: '#/components/schemas/SettlementExecutionStatus'
        reconciliationStatus:
          $ref: '#/components/schemas/SettlementReconciliationStatus'
        callbackStatus:
          $ref: '#/components/schemas/SettlementCallbackStatus'
        providerStatus:
          type: [string, 'null']
        txHash:
          type: [string, 'null']
          description: Canonical EVM transaction hash for the latest settlement execution reference.
        explorerUrl:
          type: [string, 'null']
          description: Canonical chain explorer link derived from `txHash` when available.
        externalReference:
          type: [string, 'null']
        latestEventType:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/SettlementEventType'
        latestEventDetail:
          type: [string, 'null']
        latestEventAt:
          type: [string, 'null']
          format: date-time
        callbackDeliveredAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SettlementExecutionStatus:
      type: string
      enum: [pending, accepted, queued, submitted, confirmed, failed, rejected]
    SettlementReconciliationStatus:
      type: string
      enum: [pending, matched, drift, unavailable]
    SettlementCallbackStatus:
      type: string
      enum: [pending, delivered, failed, dead_letter, disabled]
    SettlementEventType:
      type: string
      enum: [accepted, queued, submitted, confirmed, failed, rejected, reconciled, drift_detected]
...
    TradeResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/Trade'
        timestamp:
          type: string
          format: date-time
    TradeListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Trade'
        timestamp:
          type: string
          format: date-time
...
    GovernanceActionCategory:
      type: string
      enum:
        - pause
        - unpause
        - claims_pause
        - claims_unpause
        - treasury_sweep
        - treasury_payout_receiver_update
        - oracle_disable_emergency
        - oracle_update
    GovernanceActionStatus:
      type: string
      enum:
        - requested
        - submitted
        - prepared
        - broadcast_pending_verification
        - broadcast
        - pending_approvals
        - approved
        - executed
        - cancelled
        - expired
        - stale
        - failed
    GovernanceVerificationState:
      type: string
      enum: [not_required, not_started, pending, verified, failed]
    GovernanceMonitoringState:
      type: string
      enum:
        [
          not_required,
          not_started,
          pending_verification,
          pending_confirmation,
          confirmed,
          finalized,
          reverted,
          stale,
        ]
...
    GovernanceAction:
      type: object
      required:
        - actionId
        - intentKey
        - category
        - status
        - contractMethod
        - audit
        - createdAt
      properties:
        actionId:
          type: string
        intentKey:
          type: string
        proposalId:
          type: integer
          minimum: 0
          nullable: true
        category:
          $ref: '#/components/schemas/GovernanceActionCategory'
        status:
          $ref: '#/components/schemas/GovernanceActionStatus'
        contractMethod:
          type: string
        txHash:
          type: string
          nullable: true
        blockNumber:
          type: integer
          nullable: true
        tradeId:
          type: string
          nullable: true
        chainId:
          type: string
          nullable: true
        targetAddress:
          type: string
          nullable: true
        broadcastAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        executedAt:
          type: string
          format: date-time
          nullable: true
        signing:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/GovernancePreparedSigningPayload'
        finalSignerWallet:
          type: string
          nullable: true
        verificationState:
          $ref: '#/components/schemas/GovernanceVerificationState'
        verificationError:
          type: string
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        monitoringState:
          $ref: '#/components/schemas/GovernanceMonitoringState'
        audit:
          $ref: '#/components/schemas/MutationAuditRecord'
...
    TreasuryFreshness:
      type: object
      required: [source, sourceFreshAt, queriedAt, available]
      properties:
        source:
          type: string
          enum: [chain_rpc, gateway_governance_ledger]
        sourceFreshAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        queriedAt:
          type: string
          format: date-time
        available:
          type: boolean
        degradedReason:
          type: string
    TreasuryBalanceSnapshot:
      type: object
      required: [assetSymbol, raw, display]
      properties:
        assetSymbol:
          type: string
          const: USDC
        raw:
          type: string
        display:
          type: string
    TreasurySweepVisibility:
      type: object
      required: [canSweep, blockedReason]
      properties:
        canSweep:
          type: boolean
        blockedReason:
          oneOf:
            - type: 'null'
            - type: string
              enum: [claims_paused, no_claimable_balance]
    TreasuryPayoutReceiverVisibility:
      type: object
      required: [currentAddress, hasPendingUpdate, activeProposalIds]
      properties:
        currentAddress:
          type: string
        hasPendingUpdate:
          type: boolean
        activeProposalIds:
          type: array
          items:
            type: integer
            minimum: 0
    TreasuryState:
      type: object
      required:
        - paused
        - claimsPaused
        - treasuryAddress
        - treasuryPayoutAddress
        - governanceApprovalsRequired
        - governanceTimelockSeconds
        - requiredAdminCount
        - claimableBalance
        - sweepVisibility
        - payoutReceiverVisibility
      properties:
        paused:
          type: boolean
        claimsPaused:
          type: boolean
        treasuryAddress:
          type: string
        treasuryPayoutAddress:
          type: string
        governanceApprovalsRequired:
          type: integer
          minimum: 1
        governanceTimelockSeconds:
          type: integer
          minimum: 0
        requiredAdminCount:
          type: integer
          minimum: 1
        claimableBalance:
          $ref: '#/components/schemas/TreasuryBalanceSnapshot'
        sweepVisibility:
          $ref: '#/components/schemas/TreasurySweepVisibility'
        payoutReceiverVisibility:
          $ref: '#/components/schemas/TreasuryPayoutReceiverVisibility'
    TreasurySnapshotPayload:
      type: object
      required: [state, freshness]
      properties:
        state:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/TreasuryState'
        freshness:
          $ref: '#/components/schemas/TreasuryFreshness'
    TreasurySnapshotResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/TreasurySnapshotPayload'
        timestamp:
          type: string
          format: date-time
    TreasuryActionCategory:
      type: string
      enum: [treasury_sweep, treasury_payout_receiver_update]
    TreasuryActionListPayload:
      type: object
      required: [items, nextCursor, freshness]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GovernanceAction'
        nextCursor:
          anyOf:
            - type: string
            - type: 'null'
        freshness:
          $ref: '#/components/schemas/TreasuryFreshness'
    TreasuryActionListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/TreasuryActionListPayload'
        timestamp:
          type: string
          format: date-time
...
    GovernancePreparedTransactionRequest:
      type: object
      required: [chainId, to, data, value]
      properties:
        chainId:
          type: integer
        to:
          type: string
        data:
          type: string
        value:
          type: string
    GovernancePreparedSigningPayload:
      type: object
      required:
        [
          chainId,
          contractAddress,
          contractMethod,
          args,
          txRequest,
          signerWallet,
          preparedPayloadHash,
        ]
      properties:
        chainId:
          type: integer
        contractAddress:
          type: string
        contractMethod:
          type: string
        args:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
              - type: boolean
        txRequest:
          $ref: '#/components/schemas/GovernancePreparedTransactionRequest'
        signerWallet:
          type: string
        preparedPayloadHash:
          type: string
...